Need to cast from integer value to index value types
I could just make the program hold usizes, but this makes it clear when
I'm switching between values and indices, which I like.

+++ aoc-02-01/src/main.rs
@@ -9,7 +9,7 @@ fn run_intcode(mut program: Vec<i32>) -> Vec<i32> {
     while current_opcode != 99 {
         match current_opcode {
             1 => {
-                program[program[current_position + 3]] =
+                program[program[current_position + 3] as usize] =
                     program[current_position + 1] + program[current_position + 2]
             }
             other => panic!("Unknown opcode: {}", other),
